+++ test/lava.spec.php
@@ -29,4 +29,7 @@ describe('Krak Lava', function() {
     describe('Middleware', function() {
         require_once __DIR__ . '/middleware/mw.php';
     });
+    describe('Http', function() {
+        require_once __DIR__ . '/http/controller.php';
+    });
 });