+++ src/LavaPackage.php
@@ -134,7 +134,7 @@ class LavaPackage extends AbstractPackage
         $c['bootstrappers'] = new ArrayObject();
         $c['frozen'] = false;
         $c['debug'] = false;
-        $c['version'] = '0.3.1';
+        $c['version'] = '0.3.2';
         $c['name'] = 'Lava';
         $c['cli'] = PHP_SAPI === 'cli';
         $c['json_encode_options'] = 0;